Removed last lingering unattended SBCs from code. Moved Y caluclulations into super class
